requirements.txt
fastapi
uvicorn
sqlalchemy
fpdf2
pydantic
python-dotenv
psycopg2-binary
